Rename ecs_metadata_test to ecs_servicediscovery_test
** Why **

For readability

This test primarily tests ECS ServiceDiscovery, and leverages ECS
metadata in translation as part of the set up for the serviceDiscovery
feature. Additionally, this is currently the only test covering
serviceDiscovery.
