+++ test/ecs/ecs_sd/ecs_servicediscovery_test.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT
 
-package ecs_metadata
+package ecs_sd
 
 import (
 	_ "embed"
@@ -26,7 +26,7 @@ const (
 	RetryTime = 15
 	// Log group format: https://github.com/aws/amazon-cloudwatch-agent/blob/main/translator/translate/logs/metrics_collected/prometheus/ruleLogGroupName.go#L33
 	ECSLogGroupNameFormat = "/aws/ecs/containerinsights/%s/prometheus"
-	// Log stream based on job name: https://github.com/khanhntd/amazon-cloudwatch-agent/blob/ecs_metadata/integration/test/ecs/ecs_metadata/resources/extra_apps.tpl#L41
+	// Log stream based on job name: https://github.com/khanhntd/amazon-cloudwatch-agent-test/blob/main/test/ecs/ecs_sd/resources/extra_apps.tpl#41
 	LogStreamName = "prometheus-redis"
 )
