+++ test/ecs/ecs_sd/resources/ecs_prometheus.tpl
