+++ test/ecs/ecs_sd/resources/ecs_taskdef.tpl
