+++ test/ecs/ecs_sd/resources/emf_prometheus_redis_schema.json
